@@ -11,9 +11,12 @@ import sys
 import tempfile
 import base64
 
+# Fall back on using the "{python}" builder variable
+PYTHON_DEFAULT = '{python}'
+
 
 class InstallPythonReqs(Builder.Action):
-    def __init__(self, trust_hosts=False, deps=[], python=sys.executable):
+    def __init__(self, trust_hosts=False, deps=[], python=PYTHON_DEFAULT):
         self.trust_hosts = trust_hosts
         self.core = ('pip', 'setuptools', 'wheel')
         self.deps = deps
@@ -244,7 +247,7 @@ class AWSCrtPython(Builder.Action):
         parser = argparse.ArgumentParser()
         parser.add_argument('--python')
         args = parser.parse_known_args(env.args.args)[0]
-        python = args.python if args.python else sys.executable
+        python = args.python if args.python else PYTHON_DEFAULT
 
         actions = [
             InstallPythonReqs(deps=[], python=python),
