@@ -12,7 +12,7 @@ __all__ = [
     's3'
 ]
 
-__version__ = '1.0.0-dev'
+__version__ = '1.0.0.dev0'
 
 
 class NativeResource:
